Apply redux for Noticelist
