+++ front/redux/store/configureStore.js
@@ -10,14 +10,14 @@ import { composeWithDevTools } from 'redux-devtools-extension';
 //여러 리듀서 등록하기
 //sub리듀서를 하나로 합쳐줌
 import { combineReducers } from 'redux';
-import reducer from '../reducers/index';
+
 import faqReducer from '../reducers/faqReducer';
 import { createLogger } from 'redux-logger';
 import thunk from 'redux-thunk';
 import programReducer from '../reducers/programReducer';
 import reviewReducer from '../reducers/reviewReducer';
 import yearlyCalendarReducer from '../reducers/yearlyCalendarReducer'
-
+import noticeReducer from '../reducers/noticeReducer';
 
 
 const configureStore = () => {
@@ -30,7 +30,7 @@ const configureStore = () => {
   
   const store = createStore(
     combineReducers({
-      reducer,
+      noticeReducer,
       faqReducer,
       programReducer,
       reviewReducer,
